* Add libpvguiqt library. For now, it contains the model and view for the Picviz datatree ! * Fix bugs in PVCore::PVSharedPtr, when creating a shared ptr for a base pointer: each instance of PVSharedptr thus needs its own pointer value (it can be different w/ polymorphic objects), and the shared counter is *actually* incremented ! * Clean PVView/PVPlotted/PVMapped relationships. May be break some stuff, but we'll clean the rest later * Remove some OpenMP wrnings (add OpenMP everywhere !) * Fixes some test cases w/ the new interfaces of the datatree * gui-qt still needs to be fixed !

+++ libpvkernel/src/rush/PVExtractor.cpp
@@ -214,7 +214,7 @@ void PVRush::PVExtractor::dump_nraw()
 //	}
 
 	PVLOG_INFO("Nraw:\n");
-	for (int i = 0; i < picviz_min(10,_nraw.get_number_rows()); i++) {
+	for (size_t i = 0; i < picviz_min(10,_nraw.get_number_rows()); i++) {
 		PVLOG_INFO("Line %d: ", i);
 		for (int j = 0; j < _nraw.get_number_cols(); j++) {
 			std::cerr << qPrintable(_nraw.at(i,j)) << ",";
